+++ mol_dqn/chemgraph/configs/bootstrap_dqn.json
@@ -1,7 +1,7 @@
 {
   "atom_types": ["C", "O", "N"],
   "max_steps_per_episode": 40,
-  "allow_removal": true,
+  "allow_removal": false,
   "allow_no_modification": true,
   "allow_bonds_between_rings": false,
   "allowed_ring_sizes": [5, 6],
@@ -17,7 +17,7 @@
   "gamma": 0.8,
   "discount_factor": 1,
   "double_q": true,
-  "num_bootstrap_heads": 0,
+  "num_bootstrap_heads": 20,
   "prioritized": false,
   "prioritized_alpha": 0.6,
   "prioritized_beta": 0.4,
