+++ mol_dqn/chemgraph/configs/bootstrap_dqn_opt_800.json
@@ -1,22 +1,23 @@
 {
-  "atom_types": ["C", "O", "N"],
+  "atom_types": ["C", "N", "O"],
   "max_steps_per_episode": 20,
   "allow_removal": true,
   "allow_no_modification": true,
   "allow_bonds_between_rings": false,
-  "allowed_ring_sizes": [3, 5, 6],
+  "allowed_ring_sizes": [5, 6],
   "replay_buffer_size": 5000,
   "learning_rate": 0.0001,
   "learning_rate_decay_steps": 10000,
   "learning_rate_decay_rate": 0.9,
-  "num_episodes": 10000,
+  "num_episodes": 40000,
   "batch_size": 128,
   "learning_frequency": 4,
   "update_frequency": 20,
   "grad_clipping": 10,
-  "gamma": 1.0,
+  "gamma": 0.8,
+  "discount_factor": 1,
   "double_q": true,
-  "num_bootstrap_heads": 80,
+  "num_bootstrap_heads": 16,
   "prioritized": false,
   "prioritized_alpha": 0.6,
   "prioritized_beta": 0.4,
@@ -28,5 +29,5 @@
   "optimizer": "Adam",
   "batch_norm": false,
   "save_frequency": 200,
-  "max_num_checkpoints": 100
+  "max_num_checkpoints": 10
 }
