+++ mol_dqn/chemgraph/configs/bootstrap_dqn_step2.json
@@ -4,7 +4,7 @@
   "allow_removal": true,
   "allow_no_modification": true,
   "allow_bonds_between_rings": false,
-  "allowed_ring_sizes": [5, 6],
+  "allowed_ring_sizes": [5, 6, 7],
   "replay_buffer_size": 5000,
   "learning_rate": 0.0001,
   "learning_rate_decay_steps": 10000,
@@ -17,7 +17,7 @@
   "gamma": 0.8,
   "discount_factor": 1,
   "double_q": true,
-  "num_bootstrap_heads": 0,
+  "num_bootstrap_heads": 12,
   "prioritized": false,
   "prioritized_alpha": 0.6,
   "prioritized_beta": 0.4,
