+++ mol_dqn/chemgraph/configs/multi_obj_dqn.json
@@ -1,6 +1,6 @@
 {
   "atom_types": ["C", "O", "N"],
-  "max_steps_per_episode": 40,
+  "max_steps_per_episode": 20,
   "allow_removal": true,
   "allow_no_modification": true,
   "allow_bonds_between_rings": false,
@@ -9,7 +9,7 @@
   "learning_rate": 0.0001,
   "learning_rate_decay_steps": 10000,
   "learning_rate_decay_rate": 0.9,
-  "num_episodes": 5000,
+  "num_episodes": 3000,
   "batch_size": 128,
   "learning_frequency": 4,
   "update_frequency": 20,
@@ -17,7 +17,7 @@
   "gamma": 0.8,
   "discount_factor": 1,
   "double_q": true,
-  "num_bootstrap_heads": 0,
+  "num_bootstrap_heads": 20,
   "prioritized": false,
   "prioritized_alpha": 0.6,
   "prioritized_beta": 0.4,
