--- mol_dqn/chemgraph/optimize_logp.sh
@@ -1,2 +1,0 @@
-export DYLD_LIBRARY_PATH="/usr/local/cuda/lib:/usr/local/cuda/extras/CUPTI/lib"
-python optimize_logp.py --model_dir="./save/optimize_logp/2/" --hparams="./configs/bootstrap_dqn.json"
