+++ mol_dqn/chemgraph/optimize_logp_of_800_molecules_test.py
@@ -26,19 +26,18 @@ from absl import flags
 from absl.testing import flagsaver
 
 import tensorflow as tf
-from tensorflow import gfile
-from mol_dqn.chemgraph.mcts import deep_q_networks
-from mol_dqn.chemgraph.mcts.experiments import optimize_logp_of_800_molecules
-from mol_dqn.chemgraph.tensorflow import core
+from dqn import deep_q_networks
+import optimize_logp_of_800_molecules
+from dqn.tensorflow_core import core
 
 
 class OptimizeLogpOf800MoleculesTest(tf.test.TestCase):
 
   def setUp(self):
     super(OptimizeLogpOf800MoleculesTest, self).setUp()
-    self.mount_point = tempfile.mkdtemp(dir=flags.FLAGS.test_tmpdir)
+    self.mount_point = tempfile.mkdtemp()
     self.model_dir = os.path.join(self.mount_point, 'model_dir')
-    gfile.MakeDirs(self.model_dir)
+    os.mkdir(self.model_dir)
 
   def test_run(self):
     hparams = deep_q_networks.get_hparams(
