--- mol_dqn/chemgraph/py/SA_Score/README
@@ -1,9 +1,0 @@
-RDKit-based implementation of the method described in:
-
-Estimation of Synthetic Accessibility Score of Drug-like Molecules based on Molecular Complexity and Fragment Contributions
-Peter Ertl and Ansgar Schuffenhauer
-Journal of Cheminformatics 1:8 (2009)
-http://www.jcheminf.com/content/1/1/8
-
-Contribution from Peter Ertl and Greg Landrum
-
